polynomials: dtype stabilize sum_of_2d_modes and make more cupy friendly
dtype stabilize avoids fp64 infection

cupy friendliness from cast to array, numpy takes lists/tuples, cupy does not

closes #42

+++ prysm/polynomials/__init__.py
@@ -178,7 +178,7 @@ def sum_of_2d_modes(modes, weights):
 
     """
     # dot product of the 0th dim of modes and weights => weighted sum
-    return np.tensordot(modes, weights, axes=(0, 0))
+    return np.tensordot(modes, weights.astype(modes.dtype), axes=(0, 0))
 
 
 def hopkins(a, b, c, r, t, H):
